+++ src/gwt/src/org/rstudio/studio/client/workbench/commands/Commands.java
@@ -430,7 +430,12 @@ public abstract class
    public abstract AppCommand interruptTerminal();
    public abstract AppCommand sendTerminalToEditor();
    public abstract AppCommand sendToTerminal();
-    
+   public abstract AppCommand showTerminalOptions();
+   public abstract AppCommand openNewTerminalAtEditorLocation();
+   public abstract AppCommand openNewTerminalAtFilePaneLocation();
+   public abstract AppCommand sendFilenameToTerminal();
+   public abstract AppCommand setTerminalToCurrentDirectory();
+
    // Help
    public abstract AppCommand helpBack();
    public abstract AppCommand helpForward();
@@ -621,4 +626,7 @@ public abstract class
    public abstract AppCommand toggleScreenReaderSupport();
    public abstract AppCommand toggleTabKeyMovesFocus();
    public abstract AppCommand showAccessibilityOptions();
+   
+   // Internal
+   public abstract AppCommand showDomElements();
 }